Struct bpf_bindings::Jmp

pub struct Jmp {
    // some fields omitted
}

Jump Instructions

These are instruction that modify the PC register. They allow for the developer to control the flow of their program. On declaration the developer will need to state the True And False jump values. All conditionals use unsigned convention. Programs may only advance, not back track. The system MUST exit.

All Jump Instructions use reg A (Accumulator) as their base for comparisons. This means the deverlop is either comparing A to X or A to K.

The allowed Comparison operations are Equal (==), And (&), GT (>), GTE (>=). Only one comparison operator is allows per instruction.

Some Examples:

  use bpf_bindings::{Jmp,bpf_insn};

  // PC += 10
  let a: bpf_insn = Jmp::constant_jmp(10);

  //PC += if A == 82 { 10 } else { 100 };
  let b: bpf_insn = Jmp::new(10,100).eq().constant(82);

  //PC += if A & X { 12 } else { 16 };
  let c: bpf_insn = Jmp::new(12,16).and().reg_x();

Methods

impl Jmp

fn constant_jmp(k: u32) -> bpf_insn

PC += K

fn new(jt: u8, jf: u8) -> Jmp

Build a new jump

fn eq(self) -> Jmp

Compare EQ

fn and(self) -> Jmp

Compare AND

fn gt(self) -> Jmp

Compare GT

fn gte(self) -> Jmp

Compare GTE

fn constant(self, k: u32) -> bpf_insn

Against a constant

fn reg_x(self) -> bpf_insn

Against reg X (index register)