Struct bpf_bindings::LoadA

pub struct LoadA {
    // some fields omitted
}

Load Accumulator Register

These functions are use dto load the Accumulator Register (A). Examples:

  use bpf_bindings::{SliceSize,LoadA,bpf_insn};

  //A = 25
  let a: bpf_insn = LoadA::set(25);

  //A = M[100]
  let b: bpf_insn = LoadA::memory(100);

  //A = PACKETLEN
  let c: bpf_insn = LoadA::packet_len();

  //A = P[25:1] read 1 byte at the 25th index (0 based) of the packet
  let d: bpf_insn = LoadA::read_size(SliceSize::Byte).packet(25);

  //A = P[0+X:1] read 1 byte at the Index Register index of the packet
  let e: bpf_insn = LoadA::read_size(SliceSize::Byte).packet_index(0);

Methods

impl LoadA

fn new() -> LoadA

Build a new LoadA instruction

fn packet_len() -> bpf_insn

A = PACKET_LEN

fn set(k: u32) -> bpf_insn

A = K

fn memory(k: u32) -> bpf_insn

A = M[K]

fn read_size(s: SliceSize) -> LoadA

Set read size

fn packet(self, k: u32) -> bpf_insn

A = P[K:#] develoepr MUST set read size.

fn packet_index(self, k: u32) -> bpf_insn

A = P[X+K:#] developer MUST set read size