Struct bpf_bindings::Ret

pub struct Ret;

Exit Program

Object that dictates how a packet is read into the computer. This allows for the developer to trim the length of a packet, or reject a packet.

   use bpf_bindings::Ret;

   //reject the packet (server will never see packet arrive)
   let x = Ret::reject();

   //trim packet to a constant length (server will see a packet K size)
   let y = Ret::read_constant(1000);

   //trim packet to a constant length, or reject it. Depends on A contents
   let z = Ret::read_a();

Methods

impl Ret

fn accept() -> bpf_insn

This will accept the packet. Really it is telling the kernel to accept ~3.6GB length packet. So the kernel will just accept the whole packet.

fn reject() -> bpf_insn

This is reject the packet

fn read_constant(k: u32) -> bpf_insn

Trim the packet to a constant length

fn read_a() -> bpf_insn

Trim the packet to a length determined by A. Note that if A == 0 then the packet will be rejected